requirements.txt
streamlit==1.36.0
gTTS==2.5.1
pypdf==4.2.0
python-docx==1.1.2
pydub==0.25.1

